pyproject.toml
[project]
name = "avukat-ai"
version = "0.2.0"
description = "Turk Ceza Hukuku AI Asistani - RAG + Sesli Asistan + Browser Otomasyon"
requires-python = ">=3.10"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0.0",
    "websockets>=12.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",

    # Embedding
    "sentence-transformers>=3.0.0",

    # Data ingestion
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",

    # LLM clients
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-genai>=1.0.0",

    # Config
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # Utils
    "rich>=13.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
voice = ["google-genai>=1.0.0"]
browser = ["playwright>=1.40.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
]
all = ["google-genai>=1.0.0", "playwright>=1.40.0"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
